refactor: BaseViewModel 및 관련 인터페이스 파일 이동
`core:ui` 모듈 내 공통으로 사용되는 UI 베이스 클래스들의 패키지 위치를 변경하고, 이를 참조하는 모든 피처 모듈의 임포트 경로를 업데이트했습니다.

* **refactor: BaseViewModel 위치 이동 및 패키지명 변경**
    * `com.team.prezel.core.ui.BaseViewModel` -> `com.team.prezel.core.ui.base.BaseViewModel`로 이동
    * `UiState`, `UiIntent`, `UiEffect` 인터페이스를 `com.team.prezel.core.ui.base` 패키지로 관리하도록 구조 개선

* **refactor: 피처 모듈 내 참조 경로 업데이트**
    * `home`, `profile`, `login`, `history`, `splash` 등 각 피처 모듈 내 `ViewModel`, `UiState`, `UiIntent`, `UiEffect`에서 변경된 `core.ui.base` 패키지를 참조하도록 수정했습니다.
