+++ Prezel/feature/history/impl/src/main/java/com/team/prezel/feature/history/impl/contract/HistoryUiState.kt
@@ -1,7 +1,7 @@
 package com.team.prezel.feature.history.impl.contract
 
 import androidx.compose.runtime.Immutable
-import com.team.prezel.core.ui.UiState
+import com.team.prezel.core.ui.base.UiState
 import com.team.prezel.feature.history.impl.model.HistoryPageUiModel
 import kotlinx.collections.immutable.ImmutableList
