+++ Prezel/feature/home/impl/src/main/java/com/team/prezel/feature/home/impl/contract/HomeUiIntent.kt
@@ -1,6 +1,6 @@
 package com.team.prezel.feature.home.impl.contract
 
-import com.team.prezel.core.ui.UiIntent
+import com.team.prezel.core.ui.base.UiIntent
 
 internal sealed interface HomeUiIntent : UiIntent {
     data object FetchData : HomeUiIntent
