+++ Prezel/feature/login/impl/src/main/java/com/team/prezel/feature/login/impl/landing/contract/LoginUiEffect.kt
@@ -1,7 +1,7 @@
 package com.team.prezel.feature.login.impl.landing.contract
 
 import com.team.prezel.core.auth.model.AuthProvider
-import com.team.prezel.core.ui.UiEffect
+import com.team.prezel.core.ui.base.UiEffect
 import com.team.prezel.feature.login.impl.landing.model.LoginUiMessage
 
 internal sealed interface LoginUiEffect : UiEffect {
