+++ Prezel/feature/login/impl/src/main/java/com/team/prezel/feature/login/impl/landing/contract/LoginUiIntent.kt
@@ -2,7 +2,7 @@ package com.team.prezel.feature.login.impl.landing.contract
 
 import com.team.prezel.core.auth.model.AuthProvider
 import com.team.prezel.core.auth.model.AuthResult
-import com.team.prezel.core.ui.UiIntent
+import com.team.prezel.core.ui.base.UiIntent
 
 internal sealed interface LoginUiIntent : UiIntent {
     data class OnClickLogin(
