+++ Prezel/feature/login/impl/src/main/java/com/team/prezel/feature/login/impl/landing/contract/LoginUiState.kt
@@ -1,7 +1,7 @@
 package com.team.prezel.feature.login.impl.landing.contract
 
 import androidx.compose.runtime.Immutable
-import com.team.prezel.core.ui.UiState
+import com.team.prezel.core.ui.base.UiState
 
 @Immutable
 internal data class LoginUiState(
