+++ Prezel/feature/login/impl/src/main/java/com/team/prezel/feature/login/impl/terms/contract/TermsUiEffect.kt
@@ -1,6 +1,6 @@
 package com.team.prezel.feature.login.impl.terms.contract
 
-import com.team.prezel.core.ui.UiEffect
+import com.team.prezel.core.ui.base.UiEffect
 
 internal sealed interface TermsUiEffect : UiEffect {
     data object NavigateToProfile : TermsUiEffect
