+++ Prezel/feature/profile/impl/src/main/java/com/team/prezel/feature/profile/impl/contract/ProfileUiIntent.kt
@@ -1,6 +1,6 @@
 package com.team.prezel.feature.profile.impl.contract
 
-import com.team.prezel.core.ui.UiIntent
+import com.team.prezel.core.ui.base.UiIntent
 
 internal sealed interface ProfileUiIntent : UiIntent {
     data object FetchData : ProfileUiIntent
