+++ Prezel/feature/splash/impl/src/main/java/com/team/prezel/feature/splash/impl/contract/SplashUiEffect.kt
@@ -1,6 +1,6 @@
 package com.team.prezel.feature.splash.impl.contract
 
-import com.team.prezel.core.ui.UiEffect
+import com.team.prezel.core.ui.base.UiEffect
 
 sealed interface SplashUiEffect : UiEffect {
     data object NavigateToHome : SplashUiEffect
