+++ Prezel/feature/splash/impl/src/main/java/com/team/prezel/feature/splash/impl/contract/SplashUiIntent.kt
@@ -1,6 +1,6 @@
 package com.team.prezel.feature.splash.impl.contract
 
-import com.team.prezel.core.ui.UiIntent
+import com.team.prezel.core.ui.base.UiIntent
 
 sealed interface SplashUiIntent : UiIntent {
     data object CheckLoginStatus : SplashUiIntent
