+++ Prezel/feature/splash/impl/src/main/java/com/team/prezel/feature/splash/impl/contract/SplashUiState.kt
@@ -1,7 +1,7 @@
 package com.team.prezel.feature.splash.impl.contract
 
 import androidx.compose.runtime.Immutable
-import com.team.prezel.core.ui.UiState
+import com.team.prezel.core.ui.base.UiState
 
 @Immutable
 internal data class SplashUiState(
